Fix for moving textures on certain wall sides

+++ src/ray.rs
@@ -1,5 +1,5 @@
 pub struct Ray{
     pub length: f64,
     pub texture_index: u8,
-    pub texture_pos: f64
+    pub texture_pos: i32
 }
